Support SP1 compilation for hotshot-types
This PR restructures the dependencies of hotshot-types to avoid
bringing in unnecessary crates/features which cannot be compiled
for SP1/RISC-V. The main offending crates are:
* `socket2` (the RISC-V VM has no networking), which is usually
  imported via tokio with the `net` feature flag or some `alloy`
  crates
* `blst` (uses assembly, cannot be compiled for RISC-V)
* `directories` (used to load Aztec SRS from a file)

The changes made are:
* Import alloy sub-crates (e.g. `alloy-primitives`) directly, as
  needed, rather than importing the top-level `alloy` crate, which
  pulls in a lot of stuff we don't need
* Don't enable `tokio/net` unless we need it (e.g. in cliquenet)
* Pull cliquenet addr types and x25519 types into their own crate,
  so we can import the types without pulling in the whole
  networking stack
* Disable `bls12-381` Jellyfish feature which brings in `blst`
* Fetch Aztec SRS at build time instead of runtime, eliminating
  a dependency on the network and file system
