+++ Cargo.toml
@@ -12,6 +12,8 @@ members = [
     "contracts/rust/safe-tx-builder",
     "crates/builder",
     "crates/cliquenet",
+    "crates/cliquenet-addr",
+    "crates/cliquenet-x25519",
     "crates/espresso/api",
     "crates/espresso/dev-node",
     "crates/espresso/keyset",
@@ -71,6 +73,8 @@ default-members = [
     "contracts/rust/gen-vk-contract",
     "crates/builder",
     "crates/cliquenet",
+    "crates/cliquenet-addr",
+    "crates/cliquenet-x25519",
     "crates/hotshot/builder-api",
     "crates/hotshot/example-types",
     "crates/hotshot/hotshot",
@@ -134,6 +138,7 @@ alloy = { version = "1.0", features = [
     "signer-ledger",
 ] }
 alloy-compat = { path = "alloy-compat" }
+alloy-primitives = { version = "1", features = ["serde"] }
 alloy-rlp = "0.3"
 anyhow = "1.0"
 ark-bls12-381 = "0.5"
@@ -161,7 +166,9 @@ base64-bytes = "0.1"
 bigdecimal = "0.4.8"
 bimap = "0.6.3"
 bincode = "1.3.3"
-bitvec = { git = "https://github.com/EspressoSystems/bitvec-patch.git", features = ["serde"] }
+bitvec = { git = "https://github.com/EspressoSystems/bitvec-patch.git", features = [
+    "serde",
+] }
 blake3 = "1.5"
 bon = "3.8.2"
 bs58 = "0.5.1"
@@ -183,6 +190,8 @@ clap-serde-derive = "0.2.1"
 cld = "0.5"
 client = { path = "client" }
 cliquenet = { path = "crates/cliquenet", features = ["serde"] }
+cliquenet-addr = { path = "crates/cliquenet-addr", features = ["serde"] }
+cliquenet-x25519 = { path = "crates/cliquenet-x25519", features = ["serde"] }
 coarsetime = "0.1.34"
 committable = { git = "https://github.com/EspressoSystems/commit.git", features = [
     "ark-serialize",
@@ -236,7 +245,7 @@ hotshot-new-protocol = { path = "crates/hotshot/new-protocol" }
 hotshot-new-protocol-bench = { path = "crates/hotshot/new-protocol/bench" }
 hotshot-orchestrator = { path = "crates/hotshot/orchestrator" }
 hotshot-query-service = { path = "hotshot-query-service" }
-hotshot-query-service-types = { path = "hotshot-query-service-types" }
+hotshot-query-service-types = { path = "hotshot-query-service-types", default-features = false }
 hotshot-state-prover = { path = "hotshot-state-prover" }
 hotshot-task = { path = "crates/hotshot/task" }
 hotshot-task-impls = { path = "crates/hotshot/task-impls" }
@@ -284,11 +293,12 @@ jf-rescue = { git = "https://github.com/EspressoSystems/jellyfish", tag = "jf-re
 jf-rescue-compat = { package = "jf-rescue", git = "https://github.com/EspressoSystems/jellyfish", branch = "jf-relation-v0.5.0-compat", features = [
     "gadgets",
 ] }
-jf-signature = { git = "https://github.com/EspressoSystems/jellyfish", tag = "jf-signature-v0.4.0", features = [
+jf-signature = { git = "https://github.com/EspressoSystems/jellyfish", tag = "jf-signature-v0.4.2", default-features = false, features = [
     "std",
     "bls",
     "schnorr",
     "gadgets",
+    "parallel",
 ] }
 jf-signature-compat = { package = "jf-signature", git = "https://github.com/EspressoSystems/jellyfish", branch = "jf-relation-v0.5.0-compat", features = [
     "schnorr",
@@ -331,8 +341,16 @@ num_enum = "0.7"
 openssl = { version = "0.10", features = ["vendored"] }
 opentelemetry = "0.32"
 opentelemetry-appender-tracing = "0.32"
-opentelemetry-otlp = { version = "0.32", default-features = false, features = ["http-proto", "reqwest-blocking-client", "logs", "gzip-http", "experimental-http-retry"] }
-opentelemetry_sdk = { version = "0.32", default-features = false, features = ["logs"] }
+opentelemetry-otlp = { version = "0.32", default-features = false, features = [
+    "http-proto",
+    "reqwest-blocking-client",
+    "logs",
+    "gzip-http",
+    "experimental-http-retry",
+] }
+opentelemetry_sdk = { version = "0.32", default-features = false, features = [
+    "logs",
+] }
 p3-maybe-rayon = "0.2"
 parking_lot = { version = "0.12", features = ["send_guard"] }
 paste = "1.0"
@@ -399,7 +417,7 @@ test-utils = { path = "test-utils" }
 thiserror = "2"
 tide-disco = "0.9"
 time = "0.3"
-tokio = { version = "1", features = ["full"] }
+tokio = { version = "1", features = ["macros", "time"] }
 tokio-util = { version = "0.7", features = ["rt"] }
 toml = "0.8"
 tonic = { version = "0.14.5", features = ["transport"] }
@@ -472,6 +490,10 @@ opt-level = 1
 [profile.test.package.hotshot-contract-adapter]
 opt-level = 1
 [profile.test.package.cliquenet]
+opt-level = 1
+[profile.test.package.cliquenet-addr]
+opt-level = 1
+[profile.test.package.cliquenet-x25519]
 opt-level = 1
 [profile.test.package.hotshot-macros]
 opt-level = 1
