+++ crates/cliquenet-addr/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "cliquenet-addr"
+version = "0.1.0"
+edition = "2024"
+description = "IP addressing support for cliquenet"
+
+[features]
+serde = ["dep:serde"]
+
+[dependencies]
+thiserror = { workspace = true }
+
+# optional:
+serde = { workspace = true, features = ["derive"], optional = true }
+
+[dev-dependencies]
+quickcheck = { workspace = true }
