+++ crates/hotshot/types/Cargo.toml
@@ -12,18 +12,17 @@ path = "bin/mnemonic.rs"
 
 [features]
 gpu-vid = ["jf-advz/gpu-vid"]
-rlp = ["alloy-rlp"]
+rlp = ["alloy-primitives/rlp", "alloy-rlp"]
 test-srs = ["jf-advz/test-srs"]
 
 [dependencies]
-alloy = { workspace = true }
+alloy-primitives.workspace = true
 alloy-rlp = { workspace = true, optional = true }
 anyhow = { workspace = true }
 ark-bn254 = { workspace = true }
 ark-ed-on-bn254 = { workspace = true }
 ark-ff = { workspace = true }
 ark-serialize = { workspace = true }
-ark-srs = { workspace = true }
 ark-std = { workspace = true }
 async-broadcast = { workspace = true }
 async-lock = { workspace = true }
@@ -33,7 +32,8 @@ bitvec = { workspace = true }
 blake3 = { workspace = true }
 bs58 = { workspace = true }
 clap = { workspace = true }
-cliquenet = { workspace = true }
+cliquenet-addr = { workspace = true }
+cliquenet-x25519 = { workspace = true }
 committable = { workspace = true }
 derive_more = { workspace = true }
 digest = { workspace = true }
@@ -76,6 +76,12 @@ vbs = { workspace = true }
 vec1 = { workspace = true }
 versions = { workspace = true }
 vid = { workspace = true }
+
+[build-dependencies]
+ark-bn254 = { workspace = true }
+ark-serialize = { workspace = true }
+ark-srs = { workspace = true }
+jf-pcs = { workspace = true }
 
 [dev-dependencies]
 quickcheck.workspace = true
