+++ crates/hotshot/types/src/constants.rs
@@ -61,13 +61,6 @@ pub const TEST_UPGRADE_CONSTANTS: UpgradeConstants = UpgradeConstants {
     finish_offset: 20,
 };
 
-/// For `STAKE_TABLE_CAPACITY=200`, the light client prover (a.k.a. `hotshot-state-prover`)
-/// would need to generate proof for a circuit of slightly below 2^20 gates.
-/// Thus we need to support this upperbounded degree in our Structured Reference String (SRS),
-/// the `+2` is just an artifact from the jellyfish's Plonk proof system.
-#[allow(clippy::cast_possible_truncation)]
-pub const SRS_DEGREE: usize = 2u64.pow(20) as usize + 2;
-
 /// The `tide` module name for the legacy builder
 pub const LEGACY_BUILDER_MODULE: &str = "block_info";
