+++ crates/hotshot/types/src/light_client.rs
@@ -8,7 +8,7 @@
 
 use std::collections::HashMap;
 
-use alloy::primitives::{FixedBytes, U256};
+use alloy_primitives::{FixedBytes, U256};
 use ark_ed_on_bn254::EdwardsConfig as Config;
 use ark_ff::PrimeField;
 use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
