+++ crates/hotshot/types/src/traits/block_contents.rs
@@ -17,7 +17,7 @@ use std::{
     sync::Arc,
 };
 
-use alloy::primitives::FixedBytes;
+use alloy_primitives::FixedBytes;
 use async_trait::async_trait;
 use committable::{Commitment, Committable};
 use serde::{Deserialize, Serialize, de::DeserializeOwned};
