+++ crates/hotshot/types/src/vote.rs
@@ -11,7 +11,7 @@ use std::{
     marker::PhantomData,
 };
 
-use alloy::primitives::{FixedBytes, U256};
+use alloy_primitives::{FixedBytes, U256};
 use bitvec::{bitvec, vec::BitVec};
 use committable::{Commitment, Committable};
 use hotshot_utils::anytrace::*;
