+++ light-client/Cargo.toml
@@ -7,7 +7,12 @@ edition.workspace = true
 [features]
 default = ["rlp"]
 rlp = ["alloy-rlp", "hotshot-types/rlp", "espresso-types/rlp"]
-testing = ["espresso-types/testing", "bitvec", "hotshot-contract-adapter", "rand"]
+testing = [
+    "espresso-types/testing",
+    "bitvec",
+    "hotshot-contract-adapter",
+    "rand",
+]
 
 [dependencies]
 alloy = { workspace = true }
